WIP version of backend with mongo locally

+++ backend/backtester/environment.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from typing import List
-from strategies.base_strategy import Strategy
+from backtester.strategies.base_strategy import Strategy
 from datetime import date
 
 
@@ -9,5 +9,5 @@ class Environment:
     tickers: List[str]
     start_date: date
     end_date: date
-    cash: int
+    cash: float
     strategies: List[Strategy]
